+++ src/main/java/lotto/domain/Buyer.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import lotto.controller.LottoGenerator;
